requirements.txt
streamlit
pandas
openpyxl
google-generativeai
Pillow
matplotlib